+++ framework-system-starter/src/main/java/io/github/describeadmin/system/mapper/SysConfigMapper.java
@@ -1,0 +1,8 @@
+package io.github.describeadmin.system.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import io.github.describeadmin.system.entity.SysConfig;
+
+/** SysConfig Mapper。 */
+public interface SysConfigMapper extends BaseMapper<SysConfig> {
+}