+++ framework-system-starter/src/main/java/io/github/describeadmin/system/mapper/SysDictTypeMapper.java
@@ -1,0 +1,8 @@
+package io.github.describeadmin.system.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import io.github.describeadmin.system.entity.SysDictType;
+
+/** SysDictType Mapper。 */
+public interface SysDictTypeMapper extends BaseMapper<SysDictType> {
+}